When starts the container check what kind of service it is and launch the correct proccess hutch or rails

+++ api/post-up.sh
@@ -1,4 +1,8 @@
-rm -f tmp/pids/server.pid
 bundle install --quiet
-rails s -b 0.0.0.0 -p 3000
 
+if [ "$SERVICE_TYPE" = "worker" ]; then
+  sh wait-rabbit-and-connect.sh
+else
+  rm -f tmp/pids/server.pid
+  rails s -b 0.0.0.0 -p 3000
+fi
